requirements.txt
aiofiles==24.1.0
aiogram==3.17.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
attrs==25.1.0
babel==2.17.0
beautifulsoup4==4.13.3
bidict==0.23.1
bitarray==3.1.0
blinker==1.9.0
CacheControl==0.14.2
cachelib==0.13.0
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==2.1.1
ckzg==2.0.1
click==8.1.8
cryptography==44.0.0
cytoolz==1.0.1
distro==1.9.0
Django==5.1.6
djangorestframework==3.15.2
eth-account==0.13.5
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.6.1
eth-rlp==2.2.0
eth-typing==5.2.0
eth-utils==5.2.0
eth_abi==5.2.0
fastapi==0.115.8
firebase-admin==6.0.0
Flask==3.1.0
flask-babel==4.0.0
Flask-Cors==5.0.0
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-Session==0.8.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.2.2
frozenlist==1.5.0
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.20.0
google-cloud-storage==3.0.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpcio==1.70.0
grpcio-status==1.62.3
grpcio-tools==1.62.3
gunicorn==20.1.0
h11==0.14.0
hexbytes==1.3.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
magic-filter==1.0.12
Mako==1.3.9
MarkupSafe==3.0.2
msgpack==1.1.0
msgspec==0.19.0
multidict==6.1.0
nest-asyncio==1.6.0
openai==0.28.0
packaging==24.2
parsimonious==0.10.0
propcache==0.2.1
proto-plus==1.26.0
protobuf==4.25.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
pyTelegramBotAPI==4.15.4
python-dotenv==0.21.0
python-engineio==4.11.2
python-socketio==5.12.1
python-telegram-bot==21.10
pytz==2025.1
pyunormalize==16.0.0
regex==2024.11.6
requests==2.28.1
rfc3986==1.5.0
rlp==4.1.0
rsa==4.9
setuptools==75.8.2
simple-websocket==1.1.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.37
sqlparse==0.5.3
starlette==0.45.3
stripe==11.6.0
toolz==1.0.0
tqdm==4.67.1
types-requests==2.31.0.6
types-urllib3==1.26.25.14
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==1.26.20
uvicorn==0.34.0
web3==7.8.0
websockets==13.1
Werkzeug==3.1.3
wheel==0.45.1
wsproto==1.2.0
WTForms==3.2.1
yarl==1.18.3
